requirements.txt
fastapi
uvicorn[standard]
httpx
google-generativeai
